Migration/requirements.txt
azure-core==1.26.4
azure-data-tables==12.4.2
azure-cosmos==4.3.1
PyYAML==6.0.0